+++ pkg/pull/driver.go
@@ -19,12 +19,13 @@ package pull
 
 import (
 	"fmt"
+	"sync"
 
 	"github.com/rs/zerolog/log"
 )
 
 // Pull data from source following the given puller plan.
-func Pull(plan Plan, filters RowReader, source DataSource, exporter RowExporter, diagnostic TraceListener) *Error {
+func Pull(plan Plan, filters RowReader, source DataSource, exporter RowExporter, diagnostic TraceListener, workers int) *Error {
 	if err := source.Open(); err != nil {
 		return err
 	}
@@ -32,7 +33,7 @@ func Pull(plan Plan, filters RowReader, source DataSource, exporter RowExporter,
 	defer source.Close()
 
 	e := puller{source}
-	if err := e.pull(plan, filters, exporter.Export, diagnostic); err != nil {
+	if err := e.pull(plan, filters, exporter.Export, diagnostic, workers); err != nil {
 		return err
 	}
 
@@ -43,17 +44,49 @@ type puller struct {
 	datasource DataSource
 }
 
-func (e puller) pull(plan Plan, filters RowReader, export func(Row) *Error, diagnostic TraceListener) *Error {
-	for filters.Next() {
-		fileFilter := filters.Value()
+func (e puller) pull(plan Plan, filters RowReader, export func(Row) *Error, diagnostic TraceListener, workers int) *Error {
+	var wg sync.WaitGroup
+	pulls := make(chan filter)
+	errors := make(chan *Error)
+	done := make(chan bool)
 
-		initFilter := filter{plan.InitFilter().Limit(), fileFilter.Update(plan.InitFilter().Values()), plan.InitFilter().Where()}
-		if err := e.pullStep(plan.Steps().Step(0), initFilter, export, diagnostic); err != nil {
-			return err
+	for w := 1; w <= workers; w++ {
+		wg.Add(1)
+		worker := func(id int, pulls <-chan filter, errChan chan<- *Error, wg *sync.WaitGroup) {
+			defer wg.Done()
+			log.Info().Msg(fmt.Sprintf("pull: starting worker %d", id))
+
+			for initFilter := range pulls {
+				if err := e.pullStep(plan.Steps().Step(0), initFilter, export, diagnostic); err != nil {
+					errChan <- err
+				}
+			}
 		}
+		go worker(w, pulls, errors, &wg)
 	}
 
-	return filters.Error()
+	go func() {
+		for filters.Next() {
+			fileFilter := filters.Value()
+
+			initFilter := filter{plan.InitFilter().Limit(), fileFilter.Update(plan.InitFilter().Values()), plan.InitFilter().Where()}
+			pulls <- initFilter
+		}
+
+		close(pulls)
+	}()
+
+	go func() {
+		wg.Wait()
+		done <- true
+	}()
+
+	select {
+	case <-done:
+		return filters.Error()
+	case err := <-errors:
+		return err
+	}
 }
 
 func (e puller) pullStep(step Step, filter Filter, export func(Row) *Error, diagnostic TraceListener) *Error {