+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Types of changes
 ## [1.4.0]
 
 - `Added` `--parallel <int>` or `-p <int>` options to execute push with many workers.
+- `Added` `--parallel <int>` or `-p <int>` options to execute pull from filters file with many workers.
 
 ## [1.3.1]